WPF .NET application for Artist’s Information
Use the sample SQLite database (chinook.db)
An application that allows the user to browse artist’s information
The WPF .NET application provides 3 “lists” of data:
I. Shows the artists in the database.
II. Shows the albums by the artist selected in (I).
· If no artist is selected, it's empty.
III. Shows the tracks for the album selected in (II).
· If no album is selected it's empty.
New albums can be added via a button in a new window. (adds it to the database)
The album’s artist is the selected artist in (I).
The button does nothing if no artist is selected.
Data from the 3 lists are “formatted”.
For tracks, it has its the genre.